+++ planet-mitchell/src/Foldable/Partial.hs
@@ -1,11 +1,21 @@
 module Foldable.Partial
-  ( findJust
+  ( -- * Foldable
+    findJust
   , foldr1
   , maximum
   , maximumBy
   , minimum
   , minimumBy
+    -- * Bifoldable
+  , bifoldl1
+  , bifoldr1
+  , bimaximum
+  , bimaximumBy
+  , biminimum
+  , biminimumBy
   ) where
 
-import Data.Foldable (foldr1, maximum, maximumBy, minimum, minimumBy)
-import Safe.Foldable (findJust)
+import Data.Bifoldable (bifoldl1, bifoldr1, bimaximum, bimaximumBy, biminimum,
+                        biminimumBy)
+import Data.Foldable   (foldr1, maximum, maximumBy, minimum, minimumBy)
+import Safe.Foldable   (findJust)