+++ planet-mitchell/src/Functor.hs
@@ -27,13 +27,21 @@ module Functor
   , (-<<)
     -- * Extend
   , Extend(..)
+    -- * Bifunctor
+  , Bifunctor(..)
+  , Swapped(..)
+    -- * Biapply
+  , Biapply(..)
     -- * MFunctor
   , MFunctor(..)
   ) where
 
-import Control.Lens.Setter (mapped)
-import Control.Monad.Morph (MFunctor, hoist)
-import Data.Functor        (Functor(fmap, (<$)), void, ($>), (<$>))
+import Control.Lens.Iso     (Swapped(swapped))
+import Control.Lens.Setter  (mapped)
+import Control.Monad.Morph  (MFunctor, hoist)
+import Data.Bifunctor
+import Data.Bifunctor.Apply (Biapply((.>>), (<<.), (<<.>>)))
+import Data.Functor         (Functor(fmap, (<$)), void, ($>), (<$>))
 #if MIN_VERSION_base(4,11,0)
 import Data.Functor ((<&>))
 #endif