+++ planet-mitchell/src/Traversable.hs
@@ -6,16 +6,31 @@ module Traversable
   , mapAccumR
   , traverseBy
   , sequenceBy
-    -- ** Default @Functor@ implementation
-  , fmapDefault
-    -- ** Default @Foldable@ implementation
-  , foldMapDefault
     -- * Traversable1
   , Traversable1(..)
+    -- * Bitraversable
+  , Bitraversable(..)
+  , bisequence
+  , bifor
+  , bimapAccumL
+  , bimapAccumR
+    -- * Bitraversable1
+  , Bitraversable1(..)
+    -- * Default implementations
+  , fmapDefault
+  , foldMapDefault
+  , bimapDefault
+  , bifoldMapDefault
+  , bifoldMap1Default
   ) where
 
-import Data.Reflection            (sequenceBy, traverseBy)
-import Data.Semigroup.Traversable (Traversable1(sequence1, traverse1))
-import Data.Traversable           (Traversable(sequenceA, traverse),
-                                   fmapDefault, foldMapDefault, for, mapAccumL,
-                                   mapAccumR)
+import Data.Bitraversable           (Bitraversable(bitraverse),
+                                     bifoldMapDefault, bifor, bimapAccumL,
+                                     bimapAccumR, bimapDefault, bisequence)
+import Data.Reflection              (sequenceBy, traverseBy)
+import Data.Semigroup.Bitraversable (Bitraversable1(bisequence1, bitraverse1),
+                                     bifoldMap1Default)
+import Data.Semigroup.Traversable   (Traversable1(sequence1, traverse1))
+import Data.Traversable             (Traversable(sequenceA, traverse),
+                                     fmapDefault, foldMapDefault, for,
+                                     mapAccumL, mapAccumR)